+++ apps/client-web/src/client/query_subscriptions.ts
@@ -1,13 +1,20 @@
 import type { SubscriptionUpdateResult } from "@parcnet-js/client-rpc";
 import * as p from "@parcnet-js/podspec";
-import type { EntriesSchema, PODSchema, PodSpec } from "@parcnet-js/podspec";
+import {
+  type EntriesSchema,
+  type PODSchema,
+  type PodSpec,
+  podToPODData
+} from "@parcnet-js/podspec";
+import type { POD } from "@pcd/pod";
 import { EventEmitter } from "eventemitter3";
-import type { PODCollection } from "./pod_collection.js";
+import type { PODCollectionManager } from "./pod_collection_manager.js";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 interface Subscription<E extends EntriesSchema = any> {
   query: PodSpec<E>;
   serial: number;
+  collectionId: string;
 }
 
 interface QuerySubscriptionEvents {
@@ -19,7 +26,7 @@ export class QuerySubscriptions {
   private emitter = new EventEmitter<QuerySubscriptionEvents>();
   private nextSubscriptionId = 0;
 
-  constructor(private readonly pods: PODCollection) {
+  constructor(private readonly pods: PODCollectionManager) {
     this.pods.onUpdate((update) => {
       if (update.type === "delete") {
         for (const [id, sub] of this.subscriptions.entries()) {
@@ -29,8 +36,8 @@ export class QuerySubscriptions {
               "query-update",
               {
                 update: sub.query
-                  .query(this.pods.getAll())
-                  .matches.map((pod) => pod.serialize()),
+                  .query(this.pods.get(sub.collectionId).getAll())
+                  .matches.map((pod) => podToPODData(pod as POD)),
                 subscriptionId: id
               },
               sub.serial
@@ -46,8 +53,8 @@ export class QuerySubscriptions {
               "query-update",
               {
                 update: sub.query
-                  .query(this.pods.getAll())
-                  .matches.map((pod) => pod.serialize()),
+                  .query(this.pods.get(sub.collectionId).getAll())
+                  .matches.map((pod) => podToPODData(pod as POD)),
                 subscriptionId: id
               },
               sub.serial
@@ -60,12 +67,14 @@ export class QuerySubscriptions {
   }
 
   public async subscribe<E extends EntriesSchema>(
+    collectionId: string,
     query: PODSchema<E>
   ): Promise<string> {
     const subscriptionId = (this.nextSubscriptionId++).toString();
     this.subscriptions.set(subscriptionId, {
       query: p.pod(query),
-      serial: 0
+      serial: 0,
+      collectionId: collectionId
     });
 
     return subscriptionId;