+++ examples/test-app/src/apis/GPC.tsx
@@ -1,7 +1,6 @@
 import type { ProveResult } from "@parcnet-js/client-rpc";
-import type { PodspecProofRequest } from "@parcnet-js/podspec";
+import type { PODData, PodspecProofRequest } from "@parcnet-js/podspec";
 import { TicketSpec, ticketProofRequest } from "@parcnet-js/ticket-spec";
-import type { POD } from "@pcd/pod";
 import JSONBig from "json-bigint";
 import type { ReactNode } from "react";
 import { useEffect, useState } from "react";
@@ -60,7 +59,7 @@ export function GPC(): ReactNode {
   const [verified, setVerified] = useState<boolean | undefined>();
   const [identityV3, setIdentityV3] = useState<bigint | undefined>();
   const [publicKey, setPublicKey] = useState<string | undefined>();
-  const [ticket, setTicket] = useState<POD | undefined>();
+  const [ticket, setTicket] = useState<PODData | undefined>();
 
   useEffect(() => {
     void (async () => {
@@ -125,7 +124,7 @@ const request: PodspecProofRequest = {
   }
 };
 
-const gpcProof = await z.gpc.prove(request);
+const gpcProof = await z.gpc.prove({ request });
 
 `}
             </code>
@@ -133,7 +132,7 @@ const gpcProof = await z.gpc.prove(request);
           <TryIt
             onClick={async () => {
               try {
-                setProveResult(await z.gpc.prove(request));
+                setProveResult(await z.gpc.prove({ request }));
               } catch (e) {
                 console.log(e);
               }
@@ -151,7 +150,7 @@ const gpcProof = await z.gpc.prove(request);
             Verify a GPC proof like this:
             <code className="block text-xs font-base rounded-md p-2 whitespace-pre-wrap">
               {`
-const verified = await z.gpc.verify(proof, config, revealedClaims, request);
+const verified = await z.gpc.verify(proof, config, revealedClaims);
             `}
             </code>
           </p>
@@ -167,8 +166,7 @@ const verified = await z.gpc.verify(proof, config, revealedClaims, request);
                       await z.gpc.verify(
                         proveResult.proof,
                         proveResult.boundConfig,
-                        proveResult.revealedClaims,
-                        request
+                        proveResult.revealedClaims
                       )
                     );
                   }
@@ -242,7 +240,7 @@ await z.pod.insert(pod);
               });
 
               const pod = await z.pod.sign(entries);
-              await z.pod.insert(pod);
+              await z.pod.collection("Tickets").insert(pod);
               setTicket(pod);
             }}
             label="Generate Ticket"
@@ -264,19 +262,21 @@ await z.pod.insert(pod);
               {`
 const request = ticketProofRequest({
   classificationTuples: [
-    [
-      // The public key to match
-      "${publicKey}",
-      // The event ID to match
-      "${EVENT_ID}"
-    ]
+    {
+      signerPublicKey: "${publicKey}",
+      eventId: "${EVENT_ID}"
+    }
   ],
   fieldsToReveal: {
-    eventId: true
+    attendeeEmail: true
+  },
+  externalNullifier: {
+    type: "string",
+    value: "APP_SPECIFIC_NULLIFIER"
   }
 });
 
-const gpcProof = await z.gpc.prove(request);
+const gpcProof = await z.gpc.prove({ request: request.schema });
 
 `}
             </code>
@@ -286,13 +286,20 @@ const gpcProof = await z.gpc.prove(request);
               try {
                 const request = ticketProofRequest({
                   classificationTuples: [
-                    [await z.identity.getPublicKey(), EVENT_ID]
+                    {
+                      signerPublicKey: await z.identity.getPublicKey(),
+                      eventId: EVENT_ID
+                    }
                   ],
                   fieldsToReveal: {
-                    eventId: true
+                    attendeeEmail: true
+                  },
+                  externalNullifier: {
+                    type: "string",
+                    value: "APP_SPECIFIC_NULLIFIER"
                   }
                 });
-                setProveResult(await z.gpc.prove(request.schema));
+                setProveResult(await z.gpc.prove({ request: request.schema }));
               } catch (e) {
                 console.log(e);
               }