+++ packages/ticket-spec/test/ticket-utils.spec.ts
@@ -4,7 +4,7 @@ import { Identity as IdentityV3 } from "@semaphore-protocol/identity";
 import { v4 as uuidv4 } from "uuid";
 import { describe, expect, it } from "vitest";
 import { assert } from "vitest";
-import { TicketSpec } from "../src/index.js";
+import { TicketSpec, ticketProofRequest } from "../src/index.js";
 
 const identityV3 = new IdentityV3();
 const identityV4 = new IdentityV4();
@@ -51,5 +51,34 @@ describe("ticket-utils", () => {
       "attendeeName",
       "attendeeEmail"
     ]);
+  });
+
+  it("should create a ticket proof", () => {
+    const request = ticketProofRequest({
+      classificationTuples: [],
+      fieldsToReveal: {},
+      externalNullifier: { type: "string", value: "1" },
+      watermark: { type: "string", value: "1" }
+    });
+
+    const proofRequest = request.getProofRequest();
+    expect(proofRequest.externalNullifier).toEqual({
+      type: "string",
+      value: "1"
+    });
+    expect(proofRequest.watermark).toEqual({
+      type: "string",
+      value: "1"
+    });
+    expect(Object.keys(proofRequest.membershipLists)).toHaveLength(0);
+    expect(proofRequest.proofConfig.pods.ticket).toBeDefined();
+    // The owner entry is present because we specified an external nullifier
+    expect(
+      Object.keys(proofRequest.proofConfig.pods.ticket?.entries ?? {})
+    ).toHaveLength(1);
+    expect(proofRequest.proofConfig.pods.ticket?.entries?.owner).toEqual({
+      isOwnerID: "SemaphoreV4",
+      isRevealed: false
+    });
   });
 });