+++ examples/test-app/CHANGELOG.md
@@ -1,5 +1,55 @@
 # test-app
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [6e2f3e5]
+- Updated dependencies [2cc8bfc]
+- Updated dependencies [5ba27e1]
+- Updated dependencies [a1539a5]
+  - @parcnet-js/app-connector@1.0.0
+  - @parcnet-js/ticket-spec@1.0.0
+  - @parcnet-js/client-rpc@1.0.0
+  - @parcnet-js/podspec@1.0.0
+
+## 1.0.11-beta.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @parcnet-js/app-connector@1.0.0-beta.2
+
+## 1.0.11-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @parcnet-js/app-connector@1.0.0-beta.1
+  - @parcnet-js/ticket-spec@1.0.0-beta.1
+  - @parcnet-js/client-rpc@1.0.0-beta.1
+  - @parcnet-js/podspec@1.0.0-beta.1
+
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @parcnet-js/app-connector@0.0.9
+  - @parcnet-js/ticket-spec@0.0.4
+  - @parcnet-js/client-rpc@0.0.6
+  - @parcnet-js/podspec@0.0.5
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [0cc537e]
+  - @parcnet-js/podspec@0.0.4
+  - @parcnet-js/app-connector@0.0.8
+  - @parcnet-js/client-rpc@0.0.5
+  - @parcnet-js/ticket-spec@0.0.3
+
 ## 1.0.8
 
 ### Patch Changes